user-agent-api/pyproject.toml
[project]
name = "user-agent-api"
version = "0.1.0"
description = "User-Agent Interaction API for Policy Agent"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "pydantic>=2.5.0",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
]

